Draw various steps into the canvas
I want to try to play with genetic algorithms for picking the characters
to render the image. To do that I'd need to be able to compare the before
and after to determine fitness.

Right now they're all rendering into the same canvas. I think it'd clean
it up to have each step in its own canvas.